test: 하위 테스트 클래스에서만 사용되는 로직을 해당 클래스로 이동
