+++ src/test/java/wooteco/subway/admin/acceptance/LineAcceptanceTest.java
@@ -2,11 +2,17 @@ package wooteco.subway.admin.acceptance;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
 import wooteco.subway.admin.dto.LineDetailResponse;
 import wooteco.subway.admin.dto.LineResponse;
 
 import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -46,5 +52,39 @@ public class LineAcceptanceTest extends AcceptanceTest {
         // then
         List<LineResponse> linesAfterDelete = getLines();
         assertThat(linesAfterDelete.size()).isEqualTo(3);
+    }
+
+    List<LineResponse> getLines() {
+        return
+                given().when().
+                        get("/lines").
+                        then().
+                        log().all().
+                        extract().
+                        jsonPath().getList(".", LineResponse.class);
+    }
+
+    void updateLine(Long id, LocalTime startTime, LocalTime endTime) {
+        Map<String, String> params = new HashMap<>();
+        params.put("startTime", startTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
+        params.put("endTime", endTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
+        params.put("intervalTime", "10");
+
+        given().
+                body(params).
+                contentType(MediaType.APPLICATION_JSON_VALUE).
+                accept(MediaType.APPLICATION_JSON_VALUE).
+                when().
+                put("/lines/" + id).
+                then().
+                log().all().
+                statusCode(HttpStatus.OK.value());
+    }
+
+    void deleteLine(Long id) {
+        given().when().
+                delete("/lines/" + id).
+                then().
+                log().all();
     }
 }
